phase1/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.7.0
pydantic-settings==2.3.0
boto3==1.34.0
pgvector==0.3.2
sentence-transformers==3.0.0
pytest==8.2.0
httpx==0.27.0
python-jose[cryptography]==3.3.0 
